Add character escaping - Alert component
