+++ core/src/utilities/helpers/escaping-helpers.js
@@ -1,0 +1,20 @@
+// Helper methods that deal with character escaping.
+
+export const sanitizeHtml = text => {
+  return text
+    .replace(/&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;')
+    .replace(/'/g, '&#39;')
+    .replace(/&lt;br&gt;/g, '<br>');
+};
+
+export const sanitizeParam = param => {
+  return String(param)
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;')
+    .replace(/'/g, '&#39;')
+    .replace(/\//g, '&sol;');
+};
